Version 3.13.6
Added validity checking to API functions and calls.

Disabled accessor inlining (Chromium issue 134609).

Fixed bug in Math.min/max in optimized code (Chromium issue 145961).

Directly use %ObjectKeys in json stringify (Chromium issue 2312).

Fixed VS2005 build (issue 2313).

Activated fixed ES5 readonly semantics by default.

Added hardfp flag to the Makefile.

Performance and stability improvements on all platforms.

git-svn-id: http://v8.googlecode.com/svn/trunk@12466 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/arm/lithium-codegen-arm.cc
@@ -5080,8 +5080,8 @@ void LCodeGen::EmitDeepCopy(Handle<JSObject> object,
       for (int i = 0; i < elements_length; i++) {
         int64_t value = double_array->get_representation(i);
         // We only support little endian mode...
-        int32_t value_low = value & 0xFFFFFFFF;
-        int32_t value_high = value >> 32;
+        int32_t value_low = static_cast<int32_t>(value & 0xFFFFFFFF);
+        int32_t value_high = static_cast<int32_t>(value >> 32);
         int total_offset =
             elements_offset + FixedDoubleArray::OffsetOfElementAt(i);
         __ mov(r2, Operand(value_low));
